Move OSUtil to base
