+++ Base/src/main/java/io/deephaven/base/OSUtil.java
@@ -1,7 +1,7 @@
 //
 // Copyright (c) 2016-2024 Deephaven Data Labs and Patent Pending
 //
-package io.deephaven.stats.util;
+package io.deephaven.base;
 
 import org.jetbrains.annotations.NotNull;
