+++ Stats/src/main/java/io/deephaven/stats/StatsCPUCollector.java
@@ -7,7 +7,7 @@ import io.deephaven.base.verify.Assert;
 import io.deephaven.configuration.Configuration;
 import io.deephaven.internal.log.LoggerFactory;
 import io.deephaven.io.logger.Logger;
-import io.deephaven.stats.util.OSUtil;
+import io.deephaven.base.OSUtil;
 import io.deephaven.base.stats.*;
 import io.deephaven.hash.KeyedLongObjectHash;
 import io.deephaven.hash.KeyedLongObjectHashMap;
