+++ extensions/s3/src/test/java/io/deephaven/extensions/s3/S3SeekableChannelSimpleMinIOTest.java
@@ -5,7 +5,7 @@ package io.deephaven.extensions.s3;
 
 import io.deephaven.extensions.s3.S3Instructions.Builder;
 import io.deephaven.extensions.s3.testlib.SingletonContainers.MinIO;
-import io.deephaven.stats.util.OSUtil;
+import io.deephaven.base.OSUtil;
 import org.junit.jupiter.api.Assumptions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Tag;
