+++ gradle.properties
@@ -9,7 +9,7 @@
 # Re-builders who want to inherit the base version, but have their own qualifier can set -PdeephavenBaseQualifier="customQualifier": "X.Y.Z-customQualifier".
 #
 # Re-builders who want a fully custom version can set -PdeephavenBaseVersion="customVersion" -PdeephavenBaseQualifier="": "customVersion".
-deephavenBaseVersion=0.37.0
+deephavenBaseVersion=0.37.999
 deephavenBaseQualifier=SNAPSHOT
 
 #org.gradle.debug
